refactor: 로또 개수를 검증하는 책임 변경
